requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
apscheduler==3.10.4
tzlocal==5.2
python-docx==1.1.0
fpdf2==2.7.7
stripe==7.10.0
jinja2==3.1.3
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.119.0
cryptography==46.0.5
